# api/internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "milmil API",
    "version": "1.0.0",
    "description": "Self-hosted anime media server"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    }
  ],
  "paths": {
...
    "/api/v1/anime/{bangumiId}/sync-flags": {
      "patch": {
        "summary": "Update per-anime tracker sync flags",
        "description": "Set `sync_disabled` (0|1) to opt out of outgoing tracker sync, and/or `watch_status_override` to pin a canonical watch status. Omitted fields keep their current value.",
        "operationId": "updateAnimeSyncFlags",
        "tags": [
          "Anime"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "bangumiId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "sync_disabled": {
                    "type": "integer",
                    "enum": [
                      0,
                      1
                    ]
                  },
                  "watch_status_override": {
                    "type": "string",
                    "description": "Canonical watch status that overrides progress-derived inference (empty string clears the override)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Flags updated"
          },
          "400": {
            "description": "Invalid bangumiId or body"
          },
          "404": {
            "description": "Anime not found"
          }
        }
      }
    },
...
    "/api/v1/discover/anime/{id}": {
      "get": {
        "summary": "Anime detail by Bangumi ID, enriched with AniList cover and banner",
        "operationId": "getAnimeDetail",
        "tags": [
          "Discover"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Anime detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnimeDetail"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID"
          },
          "404": {
            "description": "Anime not found"
          },
          "429": {
            "description": "Upstream rate limited"
          },
          "502": {
            "description": "External service unavailable"
          }
        }
      }
    },
...
    "/api/v1/libraries/{id}/anime": {
      "get": {
        "summary": "List anime present in a library",
        "operationId": "listLibraryAnime",
        "tags": [
          "Libraries"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of anime summaries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnimeSummary"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/libraries/{id}/anime/{animeId}": {
      "delete": {
        "summary": "Delete an anime and its media files from a library",
        "operationId": "deleteLibraryAnime",
        "tags": [
          "Libraries"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "animeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Anime deleted"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/api/v1/search/anime": {
      "get": {
        "summary": "Local fuzzy search across the user's anime",
        "description": "Returns matches from the local database (titles + romanised/english aliases) with a 0..1 match score. Used by the CLI/agent surface; bypasses the audit middleware because it's read-only.",
        "operationId": "searchAnimeLocal",
        "tags": [
          "Search"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "description": "Substring matched against title / title_zh / title_en",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Match list (best score not guaranteed sorted)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "items"
                  ],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "id",
                          "title",
                          "alt_titles",
                          "score",
                          "source"
                        ],
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "bangumi_id": {
                            "type": "integer",
                            "format": "int64"
                          },
                          "anilist_id": {
                            "type": "integer",
                            "format": "int64"
                          },
                          "title": {
                            "type": "string"
                          },
                          "alt_titles": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "score": {
                            "type": "number",
                            "format": "float",
                            "description": "0..1 match score (exact=1.0, prefix=0.9, word=0.75, substring=0.6)"
                          },
                          "source": {
                            "type": "string",
                            "description": "Always `local` for this endpoint",
                            "enum": [
                              "local"
                            ]
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing q"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API token (mlml_ prefix)",
        "description": "Pass an API token obtained from POST /api/v1/auth/login or POST /api/v1/api-tokens. Tokens have the prefix `mlml_`."
      }
    },
    "schemas": {
      "AnimeDetail": {
        "allOf": [
          {
            "$ref": "#/components/schemas/AnimeSummary"
          },
          {
            "type": "object",
            "properties": {
              "synopsis": {
                "type": "string",
                "description": "Chinese synopsis from Bangumi"
              },
              "banner_image": {
                "type": "string"
              },
              "tags": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "popularity": {
                "type": "integer"
              },
              "rating": {
                "$ref": "#/components/schemas/Rating"
              }
            }
          }
        ]
      },
      "AnimeSummary": {
        "type": "object",
        "properties": {
          "bangumi_id": {
            "type": "integer"
          },
          "anilist_id": {
            "type": "integer"
          },
          "title": {
            "type": "string",
            "description": "Chinese title preferred, fallback Japanese"
          },
          "title_original": {
            "type": "string",
            "description": "Original Japanese title"
          },
          "title_en": {
            "type": "string"
          },
          "cover_image": {
            "type": "string"
          },
          "air_date": {
            "type": "string"
          },
          "episode_count": {
            "type": "integer"
          },
          "score": {
            "type": "number"
          }
        }
      },
...
      "Rating": {
        "type": "object",
        "properties": {
          "score": {
            "type": "number"
          },
          "total": {
            "type": "integer"
          }
        }
      },
...
    }
  },
...
}